+++ task-managments-main/app/models/employee.rb
@@ -8,4 +8,5 @@ class Employee < ApplicationRecord
   validates :last_name, presence: true
   validates :document, presence: true, uniqueness: true
   validates :mobile, presence: true, uniqueness: true
+  
 end
