+++ task-managments-main/app/views/tasks_employee_mailer/taskemployee.html.erb
@@ -11,7 +11,7 @@
   <h1>A new task has been created</h1>
     <p>The task has been successfully created with the following information:</p>
         <ul>
-            <li> Name:
+            <li>
         </ul>
 </body>
 </html>
