+++ task-managments-main/db/seeds.rb
@@ -1,11 +1,15 @@
+# Tag.create(name: "Programming", color: "green", enable: true )
+# Enterprise.create(nit: "123545879-1", address: "Pacochaberg", mobile: "0636853184", user_id: 2 )
+Employee.create(name: "juan", last_name: "cantillo", document: "1045274158", mobile: "3048758745", salary: 100.25, user_id: 1)
 # for i in 1..10 do
-#    Task.create(title: "Title - #{i}", description: "lorem - #{i}", start_date: "date - #{i}", end_date: "date - #{i}", status: "status - #{i}", tag_id: 1)
+#    Task.create(title: "Title - #{i}", description: "lorem - #{i}", start_date: Date.today, end_date: Date.today + 10.day , status: 1, tag_id: 1, enterprise_id: 1)
 # end
 
-p "Create roles"
-# Role.create!(name: "Admin", enable: true)
-# Role.create!(name: "Enterprise", enable: true)
-# Role.create!(name: "Employee", enable: true)
-UserRole.create(user_id: 2, role_id: 1)
-PermissionRole.create(role_id: 1, permission: "view_task")
-p "End roles"
+
+# p "Create roles"
+# # Role.create!(name: "Admin", enable: true)
+# # Role.create!(name: "Enterprise", enable: true)
+# # Role.create!(name: "Employee", enable: true)
+# UserRole.create(user_id: 2, role_id: 1)
+# PermissionRole.create(role_id: 1, permission: "view_task")
+# p "End roles"
